+++ docs/src/Fulma/Elements/Button/View.fs
@@ -6,7 +6,7 @@ open Fable.Helpers.React
 open Fable.Helpers.React.Props
 open Types
 open Fulma.Elements
-open Fulma.Grids
+open Fulma.Layouts
 
 
 let colorInteractive =