+++ docs/src/Fulma/Elements/Notification/View.fs
@@ -6,8 +6,6 @@ open Fable.Helpers.React
 open Fable.Helpers.React.Props
 open Types
 open Fulma.Elements
-open Fulma.Grids
-
 
 let basic =
     Notification.notification [ ]