+++ docs/src/Fulma/Layout/Columns/Types.fs
@@ -1,0 +1,8 @@
+module Layouts.Columns.Types
+
+type Model =
+    { Intro : string
+      BoxViewer : Viewer.Types.Model }
+
+type Msg =
+    | BoxViewerMsg of Viewer.Types.Msg