+++ docs/src/Menu/View.fs
@@ -44,15 +44,6 @@ let menuFulma currentPage subModel dispatch =
         else
             "menu-group", Fa.AngleUp
 
-    let (gridsClass, gridsIcon) =
-        if not subModel.IsGridExpanded then
-            match currentPage with
-            | Fulma (Grid _) ->
-                "menu-group is-active", Fa.AngleDown
-            | _ -> "menu-group", Fa.AngleDown
-        else
-            "menu-group", Fa.AngleUp
-
     [ Menu.label [ ] [ str "Fulma" ]
       Menu.list [ ]
         [ li [ ]
@@ -66,17 +57,9 @@ let menuFulma currentPage subModel dispatch =
                                menuItem "Hero" (Fulma (Layout Layouts.Hero)) currentPage
                                menuItem "Footer" (Fulma (Layout Layouts.Footer)) currentPage
                                menuItem "Section" (Fulma (Layout Layouts.Section)) currentPage
-                               menuItem "Level" (Fulma (Layout Layouts.Level)) currentPage ] ] ]
-      Menu.list [ ]
-        [ li [ ]
-             [ yield a [ ClassName gridsClass
-                         OnClick (fun _ -> ToggleMenu (Library.Fulma Grids) |> dispatch ) ]
-                       [ span [ ] [ str "Grids" ]
-                         Icon.faIcon [ ] gridsIcon ]
-               if subModel.IsGridExpanded then
-                    yield ul [ ]
-                             [ menuItem "Tile" (Fulma (Grid Grids.Tile)) currentPage
-                                ] ] ]
+                               menuItem "Level" (Fulma (Layout Layouts.Level)) currentPage
+                               menuItem "Tile" (Fulma (Layout Layouts.Tile)) currentPage
+                               menuItem "Columns" (Fulma (Layout Layouts.Columns)) currentPage ] ] ]
       Menu.list [ ]
         [ li [ ]
              [ yield a [ ClassName elementsClass