+++ src/Fulma/RELEASE_NOTES.md
@@ -5,6 +5,7 @@
 * Implement `Hero`
 * Implement `Section`
 * Implement `Tile`
+* BREAKING CHANGE: The `grid` module has been removed. You can find the `Columns` wrapper into `Layouts` module
 
 ### 0.2.0-beta-003